migration/requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
dnspython==2.1.0
idna==3.3
lxml==4.7.1
pymongo==4.0.1
python-dotenv==0.19.2
requests==2.27.0
soupsieve==2.3.1
urllib3==1.26.7